Restructure and tweak logging
